Update js
Update firebase js to pull into a separate file
