+++ src/firebaseConfig.js
@@ -1,26 +1,22 @@
-// import firebase from 'firebase'
-// import 'firebase/firestore'
+//import Firebase from 'firebase'
+import firebase from 'firebase'
+import 'firebase/firestore'
 
-// // firebase init goes here
-// const config = {
-//     apiKey: "AIzaSyDkKOpImjbjS2O0RhIQNJLQXx2SuYbxsfU",
-//     authDomain: "cornell-courseplan.firebaseapp.com",
-//     databaseURL: "https://cornell-courseplan.firebaseio.com",
-//     projectId: "cornell-courseplan",
-//     storageBucket: "cornell-courseplan.appspot.com",
-//     messagingSenderId: "1031551180906",
-// }
+// firebase init
+const config = {
+    apiKey: "AIzaSyDkKOpImjbjS2O0RhIQNJLQXx2SuYbxsfU",
+    authDomain: "cornell-courseplan.firebaseapp.com",
+    databaseURL: "https://cornell-courseplan.firebaseio.com",
+    projectId: "cornell-courseplan",
+    storageBucket: "cornell-courseplan.appspot.com",
+    messagingSenderId: "1031551180906",
+}
+firebase.initializeApp(config)
+let db = firebase.firestore()
 
-// firebase.initializeApp(config)
+// firebase collection
+let emailsCollection = db.collection('emails')
 
-// // firebase utils
-// const db = firebase.firestore()
-
-// // firebase collections
-// //db.collection()
-// const emailsCollection = db.ref('emails')
-
-// export {
-//     db,
-//     emailsCollection
-// }
+export default {
+    emailsCollection
+}
